core code updates - minor enhancements and format changes
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base@5528 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ trunk/src/org/hsqldb/lib/FileUtil.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2001-2015, The HSQL Development Group
+/* Copyright (c) 2001-2016, The HSQL Development Group
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -52,7 +52,7 @@ import org.hsqldb.lib.java.JavaSystem;
  * @author Campbell Burnet (boucherb@users dot sourceforge.net)
  * @author Fred Toussi (fredt@users dot sourceforge.net)
  * @author Ocke Janssen oj@openoffice.org
- * @version 2.3.3
+ * @version 2.3.4
  * @since 1.7.2
  */
 public class FileUtil implements FileAccess {
@@ -415,8 +415,10 @@ public class FileUtil implements FileAccess {
         if (tempDir.isDirectory()) {
             File[] tempList = tempDir.listFiles();
 
-            for (int i = 0; i < tempList.length; i++) {
-                tempList[i].delete();
+            if (tempList != null) {
+                for (int i = 0; i < tempList.length; i++) {
+                    tempList[i].delete();
+                }
             }
 
             tempDir.delete();
@@ -487,7 +489,7 @@ public class FileUtil implements FileAccess {
     static class DatabaseFilenameFilter implements FilenameFilter {
 
         String[]        suffixes      = new String[] {
-            ".backup", ".properties", ".script", ".data", ".log", ".lobs",
+            ".backup", ".properties", ".script", ".data", ".log", ".lobs"
         };
         String[]        extraSuffixes = new String[] {
             ".lck", ".sql.log", ".app.log"
